+++ CHANGELOG.rst
@@ -2,6 +2,9 @@ Changelog
 ---------
 
 v2.x (??? 2025)
+   - Added option `cli` to bypass command line if False, e.g. if class
+     initiated from within Python. Supersedes old logic that bypassed
+     the command line if no desc or argstr was given.
    - Cleaner commented example in documentation.
    - Documentation of font option.
    - Documentation of colors and color palettes.